make mark_beginx a size_t instead of an int
git-svn-id: svn://svn.savannah.gnu.org/nano/trunk/nano@2035 35c25a1d-7b9e-4130-9fde-d3aeb78583b8

+++ src/global.c
@@ -113,7 +113,7 @@ char *help_text;		/* The text in the help window */
 
 #ifndef NANO_SMALL
 filestruct *mark_beginbuf;	/* The begin marker buffer */
-int mark_beginx;		/* X value in the string to start */
+size_t mark_beginx;		/* X value in the string to start */
 #endif
 
 #ifndef DISABLE_OPERATINGDIR
